requirements.txt
streamlit
streamlit-chat
python-dotenv
openai==0.28